verideal-static-page/requirements.txt
## Flask app
Flask
gunicorn
requests
python-dotenv
Werkzeug
boto3